+++ components/embl-breadcrumbs-lookup/embl-breadcrumbs-lookup.js
@@ -30,17 +30,24 @@ function emblBreadcrumbsLookup(metaProperties) {
   if (emblBreadcrumbTarget.length > 1) {
     console.warn('There is more than one `[data-embl-js-breadcrumbs-lookup]` in which to insert the breadcrumbs; continuing but only the first element will be updated.');
   }
+  if (metaProperties.active == 'notSet') {
+    // @todo: we could infer the active breadcrumb if only one is passed
+    console.warn('There is no active EMBL breadcrumb specified, cannot proceed looking up breadcrumbs.');
+    return false;
+  }
 
   var majorFacets = ['who','what','where'];
 
+  // do the primairy breadcrumb first
+  emblBreadcrumbAppend(emblBreadcrumbTarget,metaProperties[metaProperties.active],metaProperties.active,'primary');
+
+  // do the non-primairy meta terms
+  // @todo: we probably shouldn't do related if there is no primairy
   for (var i = 0; i < majorFacets.length; i++) {
-    if (majorFacets[i] == metaProperties.active) {
-      emblBreadcrumbAppend(emblBreadcrumbTarget,metaProperties[majorFacets[i]],majorFacets[i],'primary');
-    } else {
-      // do the related paths
+    if (majorFacets[i] != metaProperties.active) {
       emblBreadcrumbAppend(emblBreadcrumbTarget,metaProperties[majorFacets[i]],majorFacets[i],'related');
     }
-  }
+  }  
 
   // make a 'related' label
   var relatedLabel = document.createElement("span");
@@ -238,6 +245,37 @@ function emblBreadcrumbAppend(breadcrumbTarget,termName,facet,type) {
 
     // scan through all terms and find a match, if any
     function scanTaxonomyForTerm(termName) {
+
+      // @todo: prefer UUID matches first
+
+      // We prefer profiles
+      Array.prototype.forEach.call(Object.keys(emblTaxonomy.terms), (termId) => {
+        let term = emblTaxonomy.terms[termId];
+        if (term.type == 'profile') {
+          if (term.name === termName) {
+            termObject = term;
+            return; //exit
+          }
+        }
+      }); 
+
+      // If no profile found, match other types of taxonomy entries
+      if (typeof termObject === 'undefined') {
+        Array.prototype.forEach.call(Object.keys(emblTaxonomy.terms), (termId) => {
+          let term = emblTaxonomy.terms[termId];
+          if (term.type != 'profile') { 
+            if (term.name === termName) {
+              termObject = term;
+              return; //exit
+            }
+          }
+        }); 
+      }
+
+      // If there's still no match, see if we can find a matching display name
+      // @todo: this is an easy win but creates messy matching, but maybe that's ok if you're not using UUID
+      // There's a risk of multiple "training" entries
+
       // We prefer profiles
       Array.prototype.forEach.call(Object.keys(emblTaxonomy.terms), (termId) => {
         let term = emblTaxonomy.terms[termId];
@@ -249,7 +287,7 @@ function emblBreadcrumbAppend(breadcrumbTarget,termName,facet,type) {
         }
       }); 
 
-      // If no profile found, match any entry in taxonomy
+      // If no profile found, match other types of taxonomy entries
       if (typeof termObject === 'undefined') {
         Array.prototype.forEach.call(Object.keys(emblTaxonomy.terms), (termId) => {
           let term = emblTaxonomy.terms[termId];
@@ -264,7 +302,7 @@ function emblBreadcrumbAppend(breadcrumbTarget,termName,facet,type) {
     }
 
     // don't scan for junk matches 
-    if (termName != 'notSet' && termName != '') {
+    if (termName != 'notSet' && termName != '' && termName != 'none') {
       scanTaxonomyForTerm(termName);
     }
 
@@ -286,6 +324,14 @@ function emblBreadcrumbAppend(breadcrumbTarget,termName,facet,type) {
       termObject.name_display = termName;
       termObject.uuid = 'null';
       termObject.uuid = [];
+    } else if (typeof termObject.url == 'undefined') {
+      // if entry was found but no link specified, generate a url for a search
+      var urlFacet = '';
+      if (termObject.primary != undefined) {
+        // prepare a search facet if available
+        urlFacet = '&taxonomyFacet='+termObject.primary;
+      }
+      termObject.url = 'https://www.embl.org/search/#stq='+termObject.name+urlFacet+'&origin=breadcrumbTaxonomy'; 
     }
 
     return termObject;
@@ -362,7 +408,7 @@ function emblBreadcrumbAppend(breadcrumbTarget,termName,facet,type) {
    * @param {string} [breadcrumbUrl] - a fully formed URL, or 'null' to not make a link
    */
   function formatBreadcrumb(termName,breadcrumbUrl) {
-    if (termName == '') {
+    if (termName == '' || termName == 'none') {
       // if no term, do nothing
       return '';
     }
@@ -378,13 +424,8 @@ function emblBreadcrumbAppend(breadcrumbTarget,termName,facet,type) {
   }
 
   var currentTerm = getCurrentTerm(termName);
-  // prepare a search facet if available
-  var urlFacet = '';
-  if (currentTerm.primary != undefined) {
-    urlFacet = '&taxonomyFacet='+currentTerm.primary;
-  }
   var breadcrumbId = currentTerm.uuid,
-      breadcrumbUrl = currentTerm.url || 'https://www.embl.org/search/#stq='+currentTerm.name+urlFacet+'&origin=breadcrumbTaxonomy', // if no link specified, do a search
+      breadcrumbUrl = currentTerm.url,
       breadcrumbParents = currentTerm.parents;
 
   // narrow down to the first matching element